feat: use `xcode` from CLI
NativeScript CLI exposes `$xcode`, which can be used instead of installing `xcode` in the project. As CLI's dependency is from a fork, that has not been updated for some time, change the code in a way that is compatible with it (NOTE: the code will also work with latest xcode version).